requirements.txt
runpod==1.7.0
requests==2.32.3
websocket-client==1.8.0
pynacl==1.5.0
uvloop==0.19.0 ; platform_system != "Windows"